feat: check quota in quota screen (#26)
* feat: move quota checks to CustomerQuotaScreen

* refactor: split CustomerQuotaScreen into components

* refactor: extract out content in NoQuotaCard
