+++ src/components/CustomerQuota/types.ts
@@ -1,0 +1,3 @@
+export interface CartState {
+  [category: string]: boolean | null;
+}
